+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/discover/nip23LongForm/LongFormPostViewModel.kt
@@ -346,10 +346,9 @@ class LongFormPostViewModel :
     }
 
     suspend fun sendDraftSync() {
-        if (!accountViewModel.settings.automaticallyCreateDrafts()) return
         if (message.text.toString().isBlank() && title.text.isBlank()) {
             accountViewModel.account.deleteDraftIgnoreErrors(draftTag.current)
-        } else {
+        } else if (accountViewModel.settings.automaticallyCreateDrafts()) {
             val template = createTemplate() ?: return
             accountViewModel.account.createAndSendDraftIgnoreErrors(draftTag.current, template, emptySet())
         }
