+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/discover/nip99Classifieds/NewProductViewModel.kt
@@ -325,10 +325,9 @@ open class NewProductViewModel :
     }
 
     suspend fun sendDraftSync() {
-        if (!accountViewModel.settings.automaticallyCreateDrafts()) return
         if (message.text.toString().isBlank()) {
             accountViewModel.account.deleteDraftIgnoreErrors(draftTag.current)
-        } else {
+        } else if (accountViewModel.settings.automaticallyCreateDrafts()) {
             val template = createTemplate() ?: return
             accountViewModel.account.createAndSendDraftIgnoreErrors(draftTag.current, template)
         }
