+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/notifications/publicMessages/NewPublicMessageViewModel.kt
@@ -350,10 +350,9 @@ class NewPublicMessageViewModel :
     }
 
     suspend fun sendDraftSync() {
-        if (!accountViewModel.settings.automaticallyCreateDrafts()) return
         if (message.text.toString().isBlank()) {
             accountViewModel.account.deleteDraftIgnoreErrors(draftTag.current)
-        } else {
+        } else if (accountViewModel.settings.automaticallyCreateDrafts()) {
             val broadcast = mutableSetOf<Event>()
             nip95attachments.forEach {
                 broadcast.add(it.first)
