+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/settings/ComposeSettingsScreen.kt
@@ -34,6 +34,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.vitorpamplona.amethyst.R
 import com.vitorpamplona.amethyst.model.BooleanType
+import com.vitorpamplona.amethyst.model.ConnectivityType
 import com.vitorpamplona.amethyst.model.UiSettingsFlow
 import com.vitorpamplona.amethyst.model.parseBooleanType
 import com.vitorpamplona.amethyst.ui.components.TitleExplainer
@@ -104,5 +105,45 @@ fun AutoCreateDraftsChoice(sharedPrefs: UiSettingsFlow) {
         createDraftsIndex.screenCode,
     ) {
         sharedPrefs.automaticallyCreateDrafts.tryEmit(parseBooleanType(it))
+    }
+}
+
+@Composable
+fun AiWritingHelpChoice(sharedPrefs: UiSettingsFlow) {
+    val aiIndex by sharedPrefs.automaticallyProposeAiImprovements.collectAsState()
+
+    val booleanItems =
+        persistentListOf(
+            TitleExplainer(stringRes(ConnectivityType.ALWAYS.resourceId)),
+            TitleExplainer(stringRes(ConnectivityType.NEVER.resourceId)),
+        )
+
+    SettingsRow(
+        R.string.ai_writing_setting_title,
+        R.string.ai_writing_setting_description,
+        booleanItems,
+        aiIndex.screenCode,
+    ) {
+        sharedPrefs.automaticallyProposeAiImprovements.tryEmit(parseBooleanType(it))
+    }
+}
+
+@Composable
+fun TrackedBroadcastsChoice(sharedPrefs: UiSettingsFlow) {
+    val useTrackedBroadcastsIndex by sharedPrefs.useTrackedBroadcasts.collectAsState()
+
+    val booleanItems =
+        persistentListOf(
+            TitleExplainer(stringRes(BooleanType.ALWAYS.reourceId)),
+            TitleExplainer(stringRes(BooleanType.NEVER.reourceId)),
+        )
+
+    SettingsRow(
+        R.string.tracked_broadcasts_setting_title,
+        R.string.tracked_broadcasts_setting_description,
+        booleanItems,
+        useTrackedBroadcastsIndex.screenCode,
+    ) {
+        sharedPrefs.useTrackedBroadcasts.tryEmit(parseBooleanType(it))
     }
 }
